Run all LR Downstream analysis

This function loads the single conditions LR outputs and return the LR network based analysis. It assumes that the table present the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another measure

generate_report(
  lrpaths,
  genes = NULL,
  out_path,
  sep = ",",
  threshold = 0,
  colors = NULL,
  out_file = NULL,
  report = TRUE,
  output_fmt = "html_document"
)

Arguments

lrpaths
Paths of single condition LR data

genes
list of genes to be considered in the sankey plots

out_path
output directory path

sep
character used on csv

threshold
percentage of edges to be pruned

colors
celltypes colorscheme

out_file
output file names

report
decide if a report is generated or not

output_fmt
rmarkdown render output format parameter

Value

Rmarkdown report all objects from each step

Examples

paths <- c('CTR' = system.file("extdata",
                              "ctr_nils_bm_human.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "exp_nils_bm_human.csv",
                              package = "CrossTalkeR"))
output =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')
data <- generate_report(lrpaths = paths,
                       genes = genes,
                       out_path = paste0(output,'/'),
                       threshold = 0,
                         out_file = "report.html")

#> Reading Files

#>                                                                                  CTR 
#> "/tmp/RtmpVbBFue/temp_libpath205728ee23d7/CrossTalkeR/extdata/ctr_nils_bm_human.csv" 
#> [1] 5
#>                                                                                  EXP 
#> "/tmp/RtmpVbBFue/temp_libpath205728ee23d7/CrossTalkeR/extdata/exp_nils_bm_human.csv" 
#> [1] 5

#> Create a Differential Table

#> Calculating CCI Ranking

#> EXP_x_CTR

#> Calculating GCI Ranking

#> EXP_x_CTR

#> Anntating the top Cell Genes

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> Reading KEGG annotation online:

#> Reading KEGG annotation online:

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> 'select()' returned 1:many mapping between keys and columns

#> Defining templates

#> Generating Report

#> Preparing Single Phenotype Report

#> Loading required package: igraph

#> 
#> Attaching package: 'igraph'

#> The following objects are masked from 'package:stats':
#> 
#>     decompose, spectrum

#> The following object is masked from 'package:base':
#> 
#>     union

#> Loading required package: dplyr

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:igraph':
#> 
#>     as_data_frame, groups, union

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#> Loading required package: tibble

#> 
#> Attaching package: 'tibble'

#> The following object is masked from 'package:igraph':
#> 
#>     as_data_frame

#> Loading required package: tidyr

#> 
#> Attaching package: 'tidyr'

#> The following object is masked from 'package:igraph':
#> 
#>     crossing

#> Loading required package: stringr

#> Loading required package: patchwork

#> Loading required package: factoextra

#> Loading required package: ggplot2

#> Welcome! Want to learn more? See two factoextra-related books at https://goo.gl/ve3WBa

#> Loading required package: ComplexHeatmap

#> Loading required package: grid

#> ========================================
#> ComplexHeatmap version 2.6.2
#> Bioconductor page: http://bioconductor.org/packages/ComplexHeatmap/
#> Github page: https://github.com/jokergoo/ComplexHeatmap
#> Documentation: http://jokergoo.github.io/ComplexHeatmap-reference
#> 
#> If you use it in published research, please cite:
#> Gu, Z. Complex heatmaps reveal patterns and correlations in multidimensional 
#>   genomic data. Bioinformatics 2016.
#> 
#> This message can be suppressed by:
#>   suppressPackageStartupMessages(library(ComplexHeatmap))
#> ========================================

#> Loading required package: DT

#> Printing CCI CTR

#> Printing CCI EXP

#> Loading required package: ggrepel

#> Warning: Removed 1515 rows containing missing values (geom_label_repel).

#> Warning: Removed 1409 rows containing missing values (geom_label_repel).

#> Preparing Comparative Phenotype Report

#> Loading required package: ggraph

#> Loading required package: graphlayouts

#> Loading required package: pals

#> Loading required package: extrafont

#> Registering fonts with R

#> Loading required package: colorBlindness

#> Printing CCI EXP_x_CTR

#> Warning: number of items to replace is not a multiple of replacement length

#> Warning: Removed 1636 rows containing missing values (geom_label_repel).

#> Warning: ggrepel: 87 unlabeled data points (too many overlaps). Consider increasing max.overlaps

#> Warning: ggrepel: 87 unlabeled data points (too many overlaps). Consider increasing max.overlaps

#> DONE!